Move bulk-map chain and update docs
