+++ src/chains/bulk-map/index.examples.js
