+++ src/chains/bulk-map/index.spec.js
